requirements.txt
apache-airflow
celery